backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase[async]
langchain
langchain-community
langchain-groq
langchain-anthropic
langchain-openai
